requirements.txt
pandas
numpy<2.3,>=1.24
ta
scikit-learn
lightgbm
joblib
scipy
ccxt
# ccxtpro  # Commented out - requires paid license
gspread
oauth2client
python-dotenv
PyYAML
python-telegram-bot
Flask
pytest
websocket-client
requests
solana>=0.33,<0.36
boto3
statsmodels
hvac
psutil
aiohttp>=3.9.4
cryptography
schedule
fastapi>=0.115.0
starlette>=0.37.0       # required by fastapi
httpx<0.28
uvicorn
rich
scikit-optimize
pydantic>=2.4,<2.7
numba
cachetools
prometheus-client
solders
pythclient
redis
base58
gunicorn
gevent
urllib3<2.0  # Downgrade to v1.x for LibreSSL compatibility
# statsmodels  # Duplicate entry removed

dependency-injector>=4.41.0

SQLAlchemy>=2.0.0

PyJWT>=2.8.0
